+++ tkcore/cpp_lib/main/storage2/storage.cpp
@@ -4,7 +4,7 @@
 #include "storage.h"
 #include "split_password.h"
 #include "core/key_errors.h"
-#include "utils/pass_spliter_v1.h"
+#include "storage1/pass_spliter_v1.h"
 #include "core/common.h"
 #include "salt_text/salt2.h"
 #include "salt_text/salt_base.h"
