+++ tkcore/cpp_lib/storage1/pass_spliter_v1.cpp
@@ -3,6 +3,7 @@
 //
 
 #include "pass_spliter_v1.h"
+
 // TODO  сделать непересекающиеся биты пароля
 #define PASSW_MASK 0xFF // выборка из оригинального пароля для создания пароля для дешифровки паролей
 #define LOGIN_MASK 0xA6 // выборка из оригинального пароля для создания пароля для дешифровки логинов
