+++ tkcore/cpp_lib/thekey.h
@@ -6,8 +6,5 @@
 #define THEKEY_THEKEY_H
 
 #include "core/key_core.h"
-#include "find/key_storage.h"
-#include "main/storage1/key_storage_v1.h"
-#include "core/key_errors.h"
 
 #endif //THEKEY_THEKEY_H
