+++ tkcore/term/main.cpp
@@ -5,6 +5,7 @@
 #include "utils/Interactive.h"
 #include "utils/term_utils.h"
 #include "storage_v1/storage_v1.h"
+#include "storage1/key_storage_v1.h"
 #include "storage_v2/storage_v2.h"
 #include "main/storage2/storage.h"
