+++ tkcore/term/storage_v1/storage_v1.cpp
@@ -6,6 +6,7 @@
 #include "../utils/term_utils.h"
 #include "../utils/Interactive.h"
 #include "core/common.h"
+#include "storage1/key_storage_v1.h"
 
 using namespace std;
 using namespace thekey;
